식자재 갯수 double 및 db deadlock 관련 수정

+++ src/main/java/com/MoP2022/Team10/db/model/IngredientModel.java
@@ -12,7 +12,7 @@ public class IngredientModel {
     public String img;
 
     //사용자별 정보
-    public int count;
+    public double count;
     public LocalDate expirationDate;
 
 }
